powerpc: clear gprs on interrupt routine entry
To be considered alongside (cite first patch), which clears gprs for
syscall interrupts, and wraps their handlers to permit stack-bound
parameters.

We zero GPRS r0, r2-r11, r14-r31, limiting influence of userspace values
in potential speculation gadgets. The remaining gprs are overwritten by
entry macros to interrupt handlers, irrespective of whether or not a
given handler consumes these register values.

Prior to this commit, r14-r31 are restored on a per-interrupt basis,
but now they are always restored.

These additional loads lead to an increase in the latency of interrupt
handling of ~1%.