syscall wrapper
Syscall wrapper implemented as per s390, x86, arm64, permitting register
clearing on entry, reducing caller influence on speculation within syscall routine.
Wrapper is a macro that emits symbols for handling syscall implementations with
parameters passed by stack pointer.

For platforms supporting this syscall wrapper, symbols with usual register
calling convention (`sys...`) to support syscall implementations that directly
call syscall implementations, forgoing the need to save parameters to the stack.

The Cell processor is the only unsupported platform, simply as it (FIXME)

Curiously, as the caller state is saved to the stack on syscall, we see
performance improvements under the syscall wrapper, which avoids
explicit restores when executing the syscall handler by inlining it.