powerpc: Include all arch-specific syscall prototypes
All syscall handler prototypes that are not included in either of
linux/syscalls.h or linux/compat.h are now forward declared in
asm/syscalls.h. This is required for compile-time type-checking for
syscall handlers, which is implemented later in this series.

32-bit compatibility syscall handlers are expressed in terms of types in
ppc32.h, which is exposed globally as a header so that it is accessible
by the prototype in syscalls.h

Signed-off-by: Rohan McLure <rmclure@linux.ibm.com>