powerpc: Add generic pud_pfn implementation
Page table checks which will be introduced in a future commit depend on
a pud_pfn stub being implemented. Prior to this commit, pud_pfn was
implemented with BUILD_BUG as the inline function is never included, as
its invocations in generic code are guarded by pud_devmap which returns
zero.

Export pud_pfn to all platforms which do not implement it to support
page table checks.

Signed-off-by: Rohan McLure <rmclure@linux.ibm.com>