powerpc: mm: Replace p{u,m,4}d_is_leaf with p{u,m,4}_leaf
Replace occurrences of p{u,m,4}d_is_leaf with p{u,m,4}_leaf, as the
latter is the name given to checking that a higher-level entry in
multi-level paging contains a page translation entry (pte).

Prior to this commit, the two names have both been present in the
kernel, having as far as I can tell the same exact purpose. While the
'is' in the title may better indicate that the macro/function is a
boolean returning check, the former naming scheme is standard through
all other architectures.

Signed-off-by: Rohan McLure <rmclure@linux.ibm.com>
---
V4: new patch.