randomize_kstack: Allow for architecture-specific offset storage
Limitations on various powerpc platforms mean that interrupt handlers
are written to rely only on a specific per-cpu store, known
as the paca (processor address communications area). Accesses to the
offset location must not be capable of inducing mmu faults, should
randomisation be employed on

Permit architectures to define their own storage and access logic for
the kstack offset.

Signed-off-by: Rohan McLure <rmclure@linux.ibm.com>